pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stratica"
version = "1.0.0"
description = "Stratigraphic Pattern Recognition & Paleoclimatic Temporal Reconstruction"
readme = "README.md"
authors = [
    {name = "Samir Baladi", email = "gitdeeper@gmail.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Geology",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["stratigraphy", "paleoclimatology", "physics-informed neural networks", "milankovitch cycles", "deep-time"]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "xarray>=0.20.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    
    "tensorflow>=2.8.0",
    "torch>=1.10.0",
    "transformers>=4.15.0",
    "astroquery>=0.4.0",
    "requests>=2.26.0",
    "pyyaml>=5.4.0",
    "click>=8.0.0",
    "rich>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
    "mypy>=0.900",
    "pre-commit>=2.0",
]
docs = [
    "mkdocs>=1.3.0",
    "mkdocs-material>=8.0.0",
    "mkdocstrings>=0.19.0",
    "mike>=1.0.0",
    "mkdocs-minify-plugin>=0.5.0",
    "mkdocs-git-revision-date-localized-plugin>=1.0.0",
]

[project.urls]
Homepage = "https://stratica.netlify.app"
Documentation = "https://stratica.readthedocs.io"
Repository = "https://gitlab.com/gitdeeper8/STRATICA"
DOI = "https://doi.org/10.5281/zenodo.18851076"

[project.scripts]
stratica = "stratica.cli:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
